requirements.txt
pandas==2.3.3
pymysql==1.1.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
lxml==4.9.3
pytz==2023.3
typing-extensions==4.8.0
openpyxl==3.1.2